Migration of some aspects of kaldi model to common model, fixes to transcription
