+++ elpis/engines/common/objects/model.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Tuple, Dict
 
 from elpis.engines.common.objects.dataset import Dataset
 from elpis.engines.common.objects.fsobject import FSObject
@@ -11,7 +11,6 @@ class ModelFiles(object):
     def __init__(self, basepath: Path):
         self.kaldi = PathStructure(basepath)
 
-
 class Model(FSObject):  # TODO not thread safe
     _config_file = 'model.json'
     _links = {**FSObject._links, **{"dataset": Dataset}}
@@ -21,6 +20,7 @@ class Model(FSObject):  # TODO not thread safe
         self.dataset: Optional[Dataset] = None
         self.config['dataset_name'] = None  # dataset hash has not been linked
         self.config['status'] = 'untrained'
+        self.config['stage_status'] = {}
         self.status = 'untrained'
 
     @classmethod
@@ -33,9 +33,35 @@ class Model(FSObject):  # TODO not thread safe
     def status(self):
         return self.config['status']
 
+    @property
+    def log(self):
+        return self.config['log']
+
+    @property
+    def stage_status(self):
+        return self.config['stage_status']
+
     @status.setter
     def status(self, value: str):
         self.config['status'] = value
+
+    @log.setter
+    def log(self, value: str):
+        self.config['log'] = value
+
+    @stage_status.setter
+    def stage_status(self, vals: Tuple[str, str, str]):
+        stage, status, message = vals
+        stage_status = self.config['stage_status']
+        stage_status[stage]['status'] = status
+        stage_status[stage]['message'] = message
+        self.config['stage_status'] = stage_status
+
+    def build_stage_status(self, stage_names: Dict[str, str]):
+        for stage_file, stage_name in stage_names.items():
+            stage_status = self.config['stage_status']
+            stage_status.update({stage_file: {'name': stage_name, 'status': 'ready', 'message': 'message'}})
+            self.config['stage_status'] = stage_status
 
     def link(self, dataset: Dataset, pron_dict: PronDict):
         self.dataset = dataset
